Extract helper funcs to separate file
