+++ src/schema/v2/artworksForUser/artworksForUser.ts
@@ -6,97 +6,18 @@ import {
   GraphQLString,
 } from "graphql"
 import { connectionFromArraySlice } from "graphql-relay"
-import { convertConnectionArgsToGravityArgs, extractNodes } from "lib/helpers"
+import { convertConnectionArgsToGravityArgs } from "lib/helpers"
 import { CursorPageable, pageable } from "relay-cursor-paging"
 import { createPageCursors } from "schema/v1/fields/pagination"
 import { ResolverContext } from "types/graphql"
 import { artworkConnection } from "schema/v2/artwork"
-import gql from "lib/gql"
+import {
+  getArtistAffinities,
+  getAffinityArtworks,
+  getBackfillArtworks,
+} from "./helpers"
 
-const MAX_ARTISTS = 50
 const MAX_ARTWORKS = 100
-const MIN_AFFINITY_SCORE = 0.5
-
-export const getArtistAffinities = async (
-  args: CursorPageable,
-  context: ResolverContext
-): Promise<string[]> => {
-  const { appToken, vortexGraphqlLoader, vortexGraphqlLoaderFactory } = context
-
-  const graphqlLoader =
-    vortexGraphqlLoader || vortexGraphqlLoaderFactory(appToken)
-
-  const vortexResult = await graphqlLoader({
-    query: gql`
-        query artistAffinitiesQuery {
-          artistAffinities(
-            first: ${MAX_ARTISTS}
-            minScore: ${MIN_AFFINITY_SCORE}
-            userId: "${args.userId}"
-          ) {
-            totalCount
-            edges {
-              node {
-                artistId
-                score
-              }
-            }
-          }
-        }
-      `,
-  })()
-
-  const artistIds = extractNodes(vortexResult.data?.artistAffinities).map(
-    (node: any) => node?.artistId
-  )
-
-  return artistIds
-}
-
-export const getAffinityArtworks = async (
-  artistIds: string[],
-  gravityArgs,
-  context: ResolverContext
-): Promise<any[]> => {
-  if (artistIds.length === 0) return []
-
-  const { size, offset } = gravityArgs
-  const { artworksLoader } = context
-
-  const artworkParams = {
-    artist_ids: artistIds,
-    availability: "for sale",
-    offset,
-    size,
-    sort: "-published_at",
-  }
-
-  const body = await artworksLoader(artworkParams)
-
-  return body
-}
-
-export const getBackfillArtworks = async (
-  remainingSize: number,
-  includeBackfill: boolean,
-  context: ResolverContext
-): Promise<any[]> => {
-  if (!includeBackfill || remainingSize < 1) return []
-
-  const { setItemsLoader, setsLoader } = context
-
-  const { body: setsBody } = await setsLoader({
-    key: "artwork-backfill",
-    sort: "internal_name",
-  })
-  const backfillSetId = setsBody.map((set) => set.id)[0]
-
-  if (!backfillSetId) return []
-
-  const { body: itemsBody } = await setItemsLoader(backfillSetId)
-
-  return itemsBody.slice(0, remainingSize)
-}
 
 export const artworksForUser: GraphQLFieldConfig<void, ResolverContext> = {
   description: "A connection of artworks for a user.",
